ipdb: bridge/bond IP down state assignment fix
Bridge and bond interfaces do not send IP updates in
the «down» state, and moreover, bridge interfaces are
created in the «down» state. For these cases IPDB
forces address lookup, if there were any address
changes.

FIXME: is there a way to get addr explicitly for an interfaces?
if so, it would reduce netlink traffic.

Bug-Url: https://github.com/svinota/pyroute2/issues/55